backend/orchestrator_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
arq
langchain
langgraph
langgraph-checkpoint-redis
structlog
